% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morfer.R
\name{analyze_to_array}
\alias{analyze_to_array}
\title{Retrieve  array which represent plain data return by morfeusz library}
\usage{
analyze_to_array(text_to_analyze)
}
\arguments{
\item{text_to_analyze}{Text to analyze}
}
\value{
an array in which every element represent one line from morfeusz result
}
\description{
Retrieve  array which represent plain data return by morfeusz library
}
