% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morfer.R
\name{analyzeToArray}
\alias{analyzeToArray}
\title{Retrieve  array which represent plain data return by morfeusz library}
\usage{
analyzeToArray(data)
}
\arguments{
\item{text_to_analyze}{data}
}
\value{
an array in which every element represent one line from morfeusz result
}
\description{
Retrieve  array which represent plain data return by morfeusz library
}
