% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morfer.R
\name{analyzeToCSV}
\alias{analyzeToCSV}
\title{Retrieve  CSV string which represent data return by morfeusz library}
\usage{
analyzeToCSV(morfer, data)
}
\arguments{
\item{morfer}{instance of morfer which make analyze}

\item{text_to_analyze}{data}
}
\value{
an array in which every element represent one line from morfeusz result
}
\description{
Retrieve  CSV string which represent data return by morfeusz library
}
