% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morfer.R
\name{getFileData}
\alias{getFileData}
\title{Retrieve  a Java File object for specific path}
\usage{
getFileData(filepath)
}
\arguments{
\item{filepath}{path which point to text file}
}
\value{
an File object wich represent specific file
}
\description{
Retrieve  a Java File object for specific path
}
