% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morfer.R
\name{getMorfer}
\alias{getMorfer}
\title{Retrieve a Java instance of Morfer}
\usage{
getMorfer(rLibaryPath)
}
\arguments{
\item{rLibraryPath}{path which point to text file}
}
\value{
a Java instance of Morfer
}
\description{
Retrieve a Java instance of Morfer
}
